requirements.txt
streamlit 
tensorflow 
pillow 
tensorflow-hub